requirements.txt
aws-cdk-lib
constructs
boto3
python-dotenv
